requirements.txt
# Core FastAPI and web framework
fastapi==0.116.1
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# CrewAI and LangChain for AI agents
crewai==0.1.7
langchain==0.0.351
langchain-core==0.1.23
langchain-community==0.0.20

# OpenAI and other LLM providers
openai==0.28.1
langchain-openai==0.1.0

# Data processing and analysis
pandas==2.0.3
numpy==1.24.4
yfinance==0.2.67

# Web scraping and search
duckduckgo-search==7.2.1
beautifulsoup4==4.13.4
lxml==6.0.0
requests==2.31.0

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.10.4

# Utilities
colorama==0.4.6
click==8.1.8
